# openapi.yaml
openapi: 3.0.3
info:
  title: API Viña 2026
  version: 1.0.0
  description: Documentación de endpoints.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/vina-2026/parrilla:
    get:
      summary: Obtiene la parrilla completa
      operationId: getParrilla
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          description: Unauthenticated
...